+++ libs/error-tracking/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @protolabsai/error-tracking
 
+## 0.53.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @protolabsai/types@0.60.0
+  - @protolabsai/utils@0.60.0
+
 ## 0.53.6
 
 ### Patch Changes