+++ libs/pen-parser/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @protolabsai/pen-parser
 
+## 0.53.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @protolabsai/types@0.60.0
+
 ## 0.53.6
 
 ### Patch Changes